Added extra if checks to prevent installing config files for missing commands

+++ setup-linux.sh
@@ -26,15 +26,24 @@ then
 fi
 
 # ------------------------------------------------------- #
-#               Install oh-my-zsh                         #
+#               Setup zsh and oh-my-zsh                   #
 # ------------------------------------------------------- #
-sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
-mkdir ~/.oh-my-zsh/custom/themes
-ln -s ~/.config-files/faylite.zsh-theme ~/.oh-my-zsh/custom/themes/faylite.zsh-theme
+# If zsh is installed
+if [ -z "$(command -v zsh)" ]
+then
+	# Setup zsh
+	rm ~/.zshrc
+	ln -s ~/.config-files/.zshrc ~/.zshrc
+		
+	# Setup oh-my-zsh
+	sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+	mkdir ~/.oh-my-zsh/custom/themes
+	ln -s ~/.config-files/faylite.zsh-theme ~/.oh-my-zsh/custom/themes/faylite.zsh-theme
+fi
 
 
 # ------------------------------------------------------- #
-#                      Shells                             #
+#                      Bash                               #
 # ------------------------------------------------------- #
 # Create the local_env script if it doesn't exist, sourced by zsh for env variables that shouldn't be in version control
 touch ~/.local_env.sh
@@ -43,35 +52,42 @@ touch ~/.local_env.sh
 # Load aliases into bash
 echo source ~/.config-files/.bash-aliases.sh >> ~/.bashrc
 
-# ZSH
-rm ~/.zshrc
-ln -s ~/.config-files/.zshrc ~/.zshrc
-
-
 # ------------------------------------------------------- #
 #                      Vim                                #
 # ------------------------------------------------------- #
-rm ~/.vimrc
-ln -s ~/.config-files/.vimrc ~/.vimrc
-git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
-cp ~/.config-files/vim/colors/jellybeans.vim ~/.vim/colors/.
-# Run vim command and exit
-vim +BundleInstall +qall
+if [ -z "$(command -v vim)" ]
+then
+	rm ~/.vimrc
+	ln -s ~/.config-files/.vimrc ~/.vimrc
+	git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
+	cp ~/.config-files/vim/colors/jellybeans.vim ~/.vim/colors/.
+	# Run vim command and exit
+	vim +BundleInstall +qall
+fi
 
 # ------------------------------------------------------- #
 #                      Tmux                               #
 # ------------------------------------------------------- #
-rm ~/.tmux.conf
-ln -s ~/.config-files/.tmux.conf ~/.tmux.conf
+if [ -z "$(command -v tmux)" ]
+then
+	rm ~/.tmux.conf
+	ln -s ~/.config-files/.tmux.conf ~/.tmux.conf
+fi
 
 # ------------------------------------------------------- #
 #                      Weechat                            #
 # ------------------------------------------------------- #
-rm ~/.weechat/weechat.conf
-ln -s ~/.config-files/weechat.conf ~/.weechat/weechat.conf
+if [ -z "$(command -v weechat)" ]
+then
+	rm ~/.weechat/weechat.conf
+	ln -s ~/.config-files/weechat.conf ~/.weechat/weechat.conf
+fi
 
 # ------------------------------------------------------- #
 #                   GNU Debugger                          #
 # ------------------------------------------------------- #
-rm ~/.gdbinit
-ln -s ~/.config-files/.gdbinit ~/.gdbinit
+if [ -z "$(command -v gdb)" ]
+then
+	rm ~/.gdbinit
+	ln -s ~/.config-files/.gdbinit ~/.gdbinit
+fi
